requirements.txt
Flask==2.2.3
Jinja2==3.1.2
openai==0.27.2
python-dotenv==0.21.0
